refactor(meter): unify mapping
