+++ openmeter/meter/httphandler/query.go
@@ -6,12 +6,10 @@ import (
 	"net/http"
 
 	"github.com/openmeterio/openmeter/api"
-	"github.com/openmeterio/openmeter/openmeter/customer"
 	"github.com/openmeterio/openmeter/openmeter/meter"
 	"github.com/openmeterio/openmeter/openmeter/streaming"
 	"github.com/openmeterio/openmeter/pkg/framework/commonhttp"
 	"github.com/openmeterio/openmeter/pkg/framework/transport/httptransport"
-	"github.com/samber/lo"
 )
 
 type (
@@ -194,35 +192,4 @@ func (h *handler) QueryMeterPost() QueryMeterPostHandler {
 			httptransport.WithOperationName("queryMeterPost"),
 		)...,
 	)
-}
-
-// getFilterCustomer resolves the customer IDs to customers.
-func (h *handler) getFilterCustomer(ctx context.Context, namespace string, filterCustomerIds []string) ([]customer.Customer, error) {
-	var filterCustomer []customer.Customer
-
-	if len(filterCustomerIds) == 0 {
-		return filterCustomer, nil
-	}
-
-	// List customers
-	customers, err := h.customerService.ListCustomers(ctx, customer.ListCustomersInput{
-		Namespace:   namespace,
-		CustomerIDs: filterCustomerIds,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to list customers: %w", err)
-	}
-
-	customersById := lo.KeyBy(customers.Items, func(c customer.Customer) string {
-		return c.ID
-	})
-
-	// Check if all customers are returned
-	for _, customerId := range filterCustomerIds {
-		if _, ok := customersById[customerId]; !ok {
-			return nil, fmt.Errorf("customer %s not found", customerId)
-		}
-	}
-
-	return customers.Items, nil
 }
